64-bit support for MAWs. curated data types.

+++ callbacks/MAWs_single.h
@@ -2,6 +2,7 @@
 #define MAWs_single_h
 
 
+#include "../io/io.h"
 #include "../iterator/SLT_single_string.h"
 
 
@@ -11,7 +12,7 @@
  * appended to file $filePath$;
  * @return the number of MAWs found.
  */
-unsigned int find_MAWs_single(Basic_BWT_t *BBWT, unsigned int minLength, unsigned char writeMAWs, char *filePath);
+unsigned int find_MAWs_single(Basic_BWT_t *BBWT, unsigned long minLength, unsigned char writeMAWs, char *filePath);
 
 
 #endif
